requirements.txt
aiogram==2.22.1
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.2.0
APScheduler==3.9.1
apscheduler-di==0.0.6
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
environs==9.5.0
frozenlist==1.3.1
google-api-core==2.10.0
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
httplib2==0.20.4
idna==3.4
marshmallow==3.18.0
multidict==6.0.2
oauth2client==4.1.3
oauthlib==3.2.1
packaging==21.3
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rodi==1.1.3
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
ujson==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
